Remove no_strict_path
The purpose of the no_strict_path was to allow authorized users to run
certain commands given they were in the default path. The idea was that
system programs would be harder to cause arbitrary harm than an
arbitrary program.

There were 2 main problems with this.
First is that there are plenty "dangerous" programs like "rm" already.
Secondly there are system programs that would let one subsequently run
arbitrary programs like /bin/sh. One could even get creating an start
using namespaces or bind mount to change what appears to be in the
default path.

So really the idea never made sense so we're removing it.